% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatDataWide.R
\docType{data}
\name{BatDataWide}
\alias{BatDataWide}
\title{Bat Data from Grand Teton National Park (2016-2017) in Wide-form}
\format{
A data table with 840 rows and 30 variables:
\describe{
  \item{Site}{The site from which the data were collected as a factor.}
  \item{SampleDate}{The date on which data collection began.}
  \item{Epfu}{The number of \emph{Eptesicus fuscus} call sequences positively identified by SonoBat.}
  \item{Laci}{The number of \emph{Lasiurus cinereus} call sequences positively identified by SonoBat.}
  \item{Lano}{The number of \emph{Lasionycteris noctivagans} call sequences positively identified by SonoBat.}
  \item{Myev}{The number of \emph{Myotis evotis} call sequences positively identified by SonoBat.}
  \item{Mylu}{The number of \emph{Myotis lucifugus} call sequences positively identified by SonoBat.}
  \item{Myvo}{The number of \emph{Myotis volans} call sequences positively identified by SonoBat.}
  \item{Myyu}{The number of \emph{Myotis yumanensis} call sequences positively identified by SonoBat.}
  \item{EastingUsAea}{Longitude coordinates (using the ESRI:102003 projection) for the data collection site.}
  \item{NorthingUsAea}{Latitude coordinates (using the ESRI:102003 projection) for the data collection site.}
  \item{Year}{The year in which data was collected as a factor.}
  \item{Yday}{The ordinal date of data collection.}
  \item{YdayScale}{Yday standardized using the \code{\link{scale2}} function.}
  \item{MoonPct}{The proportion of the moon that was illuminated during data collection.}
  \item{MoonScale}{MoonPct standardized using the \code{\link{scale2}} function.}
  \item{ManualForestPct}{The proportion of land surrounding the data collection site that was forest.}
  \item{ForestScale}{ManualForestPct standardized using the \code{\link{scale2}} function.}
  \item{ManualDevelPct}{The proportion of land surrounding the data collection site that was developed.}
  \item{DevelScale}{ManualDevelPct standardized using the \code{\link{scale2}} function.}
  \item{WaterDist}{The distance to the nearest water source in meters.}
  \item{WaterScale}{WaterDist standardized using the \code{\link{scale2}} function.}
  \item{Elev}{The elevation of the data collection site in meters.}
  \item{ElevScale}{Elev standardized using the \code{\link{scale2}} function.}
  \item{BrightCount}{The sum of brightness scores for lights surrounding the site of data collection.}
  \item{BrightScale}{BrightCount standardized using the \code{\link{scale2}} function.}
  \item{PropCool}{The proportion of lights with a subjectively classified cool hue surrounding the site of data collection.}
  \item{CoolScale}{PropCool standardized using the \code{\link{scale2}} function.}
  \item{StrWeight}{The summed inverse of the square root of distances (in meters) to porous buildings surrounding the site of data collection.}
  \item{StrScale}{StrWeight standardized using the \code{\link{scale2}} function.}
}
}
\usage{
data(BatDataWide)
}
\description{
The dataset used for the case study in Cole et al., 2021 in wide-form.
}
\keyword{datasets}
