% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistFitWide.R
\name{DistFitWide}
\alias{DistFitWide}
\title{Visually examine the mean-variance relationship of data}
\usage{
DistFitWide(Splitters, Data, GroupList, ThemeBlack = T)
}
\arguments{
\item{Splitters}{A character vector of column names to aggregate data by.}

\item{Data}{The unquoted name of the data frame or data table containing the raw data models are based on.}

\item{GroupList}{A character vector of column names representing groups that define model membership (e.g., species, demographics).}

\item{ThemeBlack}{A Boolean value indicating whether the plots generated should include \code{\link{theme_nocturnal}}
(T) or not (F).}
}
\value{
This function produces a mean/variance plot with NB1, NB2, and Poisson family function lines for
each group in the vector supplied to the \code{GroupList} argument. Each resultant plot is named with
the group name followed by "DistPlot" (e.g., the plot associated with a group named "Epfu" would be named
"EpfuDistPlot".)
}
\description{
This function plots data by group, and overlays lines representing three family functions that are
commonly used for count data: NB1, NB2, and Poisson. For more information on this, please reference 
\href{https://groups.nceas.ucsb.edu/non-linear-modeling/projects/owls/WRITEUP/owls.pdf/@@download}{this document.}
}
\examples{
data("BatDataWide", package = "EcoCountHelper")

DistFitWide(c("Site", "Year"),
                BatDataWide, c("Myev", "Epfu"))
                
EpfuDistPlot
MyevDistPlot

}
\seealso{
\code{\link{DistFitLong}}
}
