% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DumbGrid.R
\name{DumbGrid}
\alias{DumbGrid}
\title{A not-so-smart plot arranger}
\usage{
DumbGrid(..., Ncols, FirstColWidth = 1, SharedLegend = F)
}
\arguments{
\item{...}{Unquoted object names of plots to be included in the resultant grid of plots.}

\item{Ncols}{A number indicating the number of columns that should be used in the plot grid.}

\item{FirstColWidth}{A number indicating the relative width of the first column of plots compared to all
other columns. A value of 1 sets the left-most column at the same width as all others.}

\item{SharedLegend}{A Boolean value indicating whether plots should share a legend (T), or have their own
legend (F).}
}
\value{
This function produces a grid of plots constructed from the plots supplied to the \code{...}
argument. Plots that are not in the left-most columns will not have y-axis labels.
}
\description{
This function is a wrapper for \code{\link[cowplot]{plot_grid}} that removes unnecessary axis labels for
plots that are not along the left side of the plot. While it was initially intended to incorporate
a "smart-adjust" feature that would scale all plots without y-axis labels to be the same width as the plot
areas with y-axis labels, this function requires manual resizing of the furthest-left column of plots to
make them the same size as the others, hence the name DumbGrid rather than SmartGrid.
}
\examples{
data("ExPlot", package = "EcoCountHelper")
TestGrid <- DumbGrid(ExPlot, ExPlot, ExPlot, ExPlot,
                      Ncols = 2)
TestGrid #View the plot in a separate window

#Adjusting the first column width to make plot sizes equal

TestGrid2 <- DumbGrid(ExPlot, ExPlot, ExPlot, ExPlot,
                       Ncols = 2, FirstColWidth = 1.35)
TestGrid2 #View the plot in a separate window

}
