% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectsPlotter.R
\name{EffectsPlotter}
\alias{EffectsPlotter}
\title{Visualize model effects}
\usage{
EffectsPlotter(
  TopMods,
  ParamLabs = NULL,
  TopModCol = NULL,
  ConfInts = c(90, 80),
  Scaled = T,
  ThemeBlack = T
)
}
\arguments{
\item{TopMods}{This argument can accept multiple input types, and is used to specify which models effects plots
should be generated for. If an effects plot is desired for a single model, users can supply the unquoted model
object name. If effects plots for multiple models are desired, users can supply either the unquoted name
of a character vector containing model object names, or the name of a data table or data frame containing a
character vector of model object names. If using the latter option, be sure to provide a value to the
\code{TopModCol} argument.}

\item{ParamLabs}{A character vector of labels for each model estimate. Labels must be supplied for estimates in the
same order as the vector returned by \code{row.names(summary(x)$coeff$cond)}, where \code{x} is a model
object name.}

\item{TopModCol}{A character string specifying the name of the column containing model object names in a
data frame or data table. This argument should only be used if \code{TopMods} is supplied with the name of
a data frame or data table such as the top model table returned by \code{\link{ModelCompare}}.}

\item{ConfInts}{A numeric vector specifying up to three condifence intervals as percent values from largest
to smallest (e.g., for 80\%, 90\%, and 95\% confidence intervals, supply \code{ConfInts = c(95, 90, 80)}).}

\item{Scaled}{A Boolean value indicating whether model estimates are scaled (T) or not (F). When \code{Scaled = T},
there will be a dotted \code{\link[ggplot2]{geom_abline}} where estimates = 0.}

\item{ThemeBlack}{A Boolean value indicating whether \code{\link{theme_nocturnal}} should be applied to plots
(T) or not (F).}
}
\value{
This function generates effects plots using \code{\link[ggplot2]{geom_point}} for point estimates and
\code{\link[ggplot2]{geom_errorbar}} for confidence intervals. Resultant plots are named with the model name
followed by "EffectsPlot" (e.g., for a model named "EpfuNb2", the effects plot would be named "EpfuNb2EffectsPlot").
}
\description{
This function produces effects plots for the conditional model of model objects produced by
\code{\link[glmmTMB]{glmmTMB}}. Effects can be plotted for multiple models in a single function call.
These effects plots are based on those found in 
\href{https://groups.nceas.ucsb.edu/non-linear-modeling/projects/owls/WRITEUP/owls.pdf/@@download}{this document.}
}
\section{Details}{

Because all models specified in a call to this function share estimate labels, all models in a a given call
to this function must have the same model formula.
}

\examples{
data("EpfuNb2Long", "MyevNb2Long", package = "EcoCountHelper")
Labels <- letters[1:12]

#Effects plot for a single model

EffectsPlotter(EpfuNb2Long, Labels)
EpfuNb2LongEffectsPlot

#Effects plot for multiple models specified
#in a character vector.

Mods <- c("EpfuNb2Long", "MyevNb2Long")
EffectsPlotter(Mods, Labels)
EpfuNb2LongEffectsPlot
MyevNb2LongEffectsPlot

#Effects plot for multiple models specified
#in a data frame

ModTable <- data.frame(Species = c("Epfu", "Myev"),
                        Mods = c("EpfuNb2Long", "MyevNb2Long"))
                        
EffectsPlotter(ModTable, Labels, "Mods")

EpfuNb2LongEffectsPlot
MyevNb2LongEffectsPlot

}
