% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epfu_Nb2.R
\docType{data}
\name{Epfu_Nb2}
\alias{Epfu_Nb2}
\title{Eptesicus fuscus model using nbinom2() family (Wide-form data).}
\format{
A `glmmTMB` class model generated using the BatDataWide data set, the `nbinom2()` 
family function, and the general conditional model structure 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Epfu_Nb2)
}
\description{
The model object named 'Epfu_Nb2' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
