% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epfu_Poi.R
\docType{data}
\name{Epfu_Poi}
\alias{Epfu_Poi}
\title{Eptesicus fuscus model using poisson() family (Wide-form data).}
\format{
A `glmmTMB` class model generated using the BatDataWide data set, the `poisson()` 
family function, and the general conditional model structure 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Epfu_Poi)
}
\description{
The model object named 'Epfu_Poi' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
