% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epfu_ZiPoi_Long.R
\docType{data}
\name{Epfu_ZiPoi_Long}
\alias{Epfu_ZiPoi_Long}
\title{Eptesicus fuscus model using poisson() family and zero-inflated formula (Long-form data).}
\format{
A `glmmTMB` class model generated using the BatDataLong data set, the `poisson()` 
family function, and the general conditional model structure and the zero-inflated model 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Epfu_ZiPoi_Long)
}
\description{
The model object named 'Epfu_ZiPoi_Long' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
