% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lano_Nb2_Long.R
\docType{data}
\name{Lano_Nb2_Long}
\alias{Lano_Nb2_Long}
\title{Lasionycteris noctivagans model using nbinom2() family (Long-form data).}
\format{
A `glmmTMB` class model generated using the BatDataLong data set, the `nbinom2()` 
family function, and the general conditional model structure 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Lano_Nb2_Long)
}
\description{
The model object named 'Lano_Nb2_Long' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
