% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lano_Poi.R
\docType{data}
\name{Lano_Poi}
\alias{Lano_Poi}
\title{Lasionycteris noctivagans model using poisson() family (Wide-form data).}
\format{
A `glmmTMB` class model generated using the BatDataWide data set, the `poisson()` 
family function, and the general conditional model structure 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Lano_Poi)
}
\description{
The model object named 'Lano_Poi' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
