% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lano_Poi_Long.R
\docType{data}
\name{Lano_Poi_Long}
\alias{Lano_Poi_Long}
\title{Lasionycteris noctivagans model using poisson() family (Long-form data).}
\format{
A `glmmTMB` class model generated using the BatDataLong data set, the `poisson()` 
family function, and the general conditional model structure 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Lano_Poi_Long)
}
\description{
The model object named 'Lano_Poi_Long' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
