% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCompare.R
\name{ModelCompare}
\alias{ModelCompare}
\title{Model selection tool for multiple groups}
\usage{
ModelCompare(Groups, TopModOutName)
}
\arguments{
\item{Groups}{A character vector of groups (e.g. species, demographics), each with two or more candidate
models to be compared. Because the values supplied for this argument are essentially used as
regular expression text patterns to search the global environment, model object names must contain a
value supplied to this argument to be included in the comparisons.}

\item{TopModOutName}{An unquoted string used as the object name for the resultant data table showing
the top model for each group supplied to the \code{Groups} argument.}
}
\value{
This function produces n+1 data tables, with n being the number of groups supplied to the
\code{Groups} argument. A data table detailing each model in a group and its AIC value is produed for each
group. Each of these data tables is named with the group name followed by "AIC" (e.g., the data table
associated with a group named "Epfu" would be named "EpfuAIC"). An additional data table is produced
detailing the top model for each group. The data table of top models is named with the value supplied
to the \code{TopModOutName} argument.
}
\description{
This function allows users to input a list/vector of groups (taxonomic or otherwise), each of which
has multiple candidate models generated with \code{\link[glmmTMB]{glmmTMB}}, and quickly attain AIC tables for each group. Group names supplied to the
\code{Groups} argument are used as text patterns supplied to \code{\link[base]{ls}} as \code{pattern}
argument values, and model objects in the global environment containing the supplied group names are
compared via \code{\link[stats]{AIC}}.
}
\examples{
data("Epfu_Nb1_Long", "Epfu_Nb2_Long", "Myev_Nb1_Long",
      "Myev_Nb2_Long", package = "EcoCountHelper")
      
Species <- c("Epfu", "Myev")
ModelCompare(Species, TestCompare)

EpfuAIC
MyevAIC
TestCompare

}
