% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myev_Nb1_Long.R
\docType{data}
\name{Myev_Nb1_Long}
\alias{Myev_Nb1_Long}
\title{Myotis evotis model using nbinom1() family (Long-form data).}
\format{
A `glmmTMB` class model generated using the BatDataLong data set, the `nbinom1()` 
family function, and the general conditional model structure 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Myev_Nb1_Long)
}
\description{
The model object named 'Myev_Nb1_Long' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
