% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myev_Poi.R
\docType{data}
\name{Myev_Poi}
\alias{Myev_Poi}
\title{Myotis evotis model using poisson() family (Wide-form data).}
\format{
A `glmmTMB` class model generated using the BatDataWide data set, the `poisson()` 
family function, and the general conditional model structure 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Myev_Poi)
}
\description{
The model object named 'Myev_Poi' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
