% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mylu_Nb1.R
\docType{data}
\name{Mylu_Nb1}
\alias{Mylu_Nb1}
\title{Myotis lucifugus model using nbinom1() family (Wide-form data).}
\format{
A `glmmTMB` class model generated using the BatDataWide data set, the `nbinom1()` 
family function, and the general conditional model structure 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Mylu_Nb1)
}
\description{
The model object named 'Mylu_Nb1' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
