% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mylu_Poi.R
\docType{data}
\name{Mylu_Poi}
\alias{Mylu_Poi}
\title{Myotis lucifugus model using poisson() family (Wide-form data).}
\format{
A `glmmTMB` class model generated using the BatDataWide data set, the `poisson()` 
family function, and the general conditional model structure 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Mylu_Poi)
}
\description{
The model object named 'Mylu_Poi' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
