% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myvo_Nb2.R
\docType{data}
\name{Myvo_Nb2}
\alias{Myvo_Nb2}
\title{Myotis volans model using nbinom2() family (Wide-form data).}
\format{
A `glmmTMB` class model generated using the BatDataWide data set, the `nbinom2()` 
family function, and the general conditional model structure 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Myvo_Nb2)
}
\description{
The model object named 'Myvo_Nb2' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
