% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myvo_Poi.R
\docType{data}
\name{Myvo_Poi}
\alias{Myvo_Poi}
\title{Myotis volans model using poisson() family (Wide-form data).}
\format{
A `glmmTMB` class model generated using the BatDataWide data set, the `poisson()` 
family function, and the general conditional model structure 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Myvo_Poi)
}
\description{
The model object named 'Myvo_Poi' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
