% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myvo_ZiPoi.R
\docType{data}
\name{Myvo_ZiPoi}
\alias{Myvo_ZiPoi}
\title{Myotis volans model using poisson() family and zero-inflated formula (Wide-form data).}
\format{
A `glmmTMB` class model generated using the BatDataWide data set, the `poisson()` 
family function, and the general conditional model structure and the zero-inflated model 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Myvo_ZiPoi)
}
\description{
The model object named 'Myvo_ZiPoi' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
