% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myvo_ZiPoi_Long.R
\docType{data}
\name{Myvo_ZiPoi_Long}
\alias{Myvo_ZiPoi_Long}
\title{Myotis volans model using poisson() family and zero-inflated formula (Long-form data).}
\format{
A `glmmTMB` class model generated using the BatDataLong data set, the `poisson()` 
family function, and the general conditional model structure and the zero-inflated model 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Myvo_ZiPoi_Long)
}
\description{
The model object named 'Myvo_ZiPoi_Long' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
