% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myyu_Nb2.R
\docType{data}
\name{Myyu_Nb2}
\alias{Myyu_Nb2}
\title{Myotis yumanensis model using nbinom2() family (Wide-form data).}
\format{
A `glmmTMB` class model generated using the BatDataWide data set, the `nbinom2()` 
family function, and the general conditional model structure 
specified in the 'EcoCountHelperExample' vignette.
}
\usage{
data(Myyu_Nb2)
}
\description{
The model object named 'Myyu_Nb2' generated according to the instructions in the 
'EcoCountHelperExample' vignette.
}
\keyword{datasets}
