% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealEffectText.R
\name{RealEffectText}
\alias{RealEffectText}
\title{Your personal glmmTMB model interpreter}
\usage{
RealEffectText(
  Model,
  Predictor,
  UnitChange = 1,
  ConfInt = 95,
  ScaleSds = NULL,
  PredVect = NULL,
  UnitSymb = "unit"
)
}
\arguments{
\item{Model}{The unquoted name of the model object of interest.}

\item{Predictor}{A character string representing the predictor of interest. Appropriate values can be found by 
entering \code{row.names(summary(x)$coeff$cond)} in the console, where x is the model of interest.}

\item{UnitChange}{A number indicating the unscaled change in a predictor for which an effect should be returned.}

\item{ConfInt}{A number indicating the desired confidence interval as a percent value.}

\item{ScaleSds}{A number indicating how many standard deviations a predictor has been scaled by.}

\item{PredVect}{The unquoted name of a vector containing the raw data for the predictor of interest.}

\item{UnitSymb}{A character string representing the units of the precictor of interest.}
}
\value{
This function returns a sentence describing the effect of a specified change in a predictor on the
response variable.
}
\description{
This function generates a sentence describing the effect of a model predictor on the response variable.
If model estimates are scaled and use a log link, this function can give "unscaled" effects of a predictor
on the response variable.
}
\examples{
#The effect of a 10\% change increase in
#moon face illumination on Epfu activity

data("EpfuNb2Long", "BatDataLong", package = "EcoCountHelper")

RealEffectText(EpfuNb2Long, "scale2(MoonPct)",
                0.1, ScaleSds = 2,
                PredVect = BatDataLong$MoonPct)

}
