% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale2.R
\name{scale2}
\alias{scale2}
\title{Scale data by two standard deviations}
\usage{
scale2(x)
}
\arguments{
\item{x}{A vector to be scaled.}
}
\description{
This function is very similar to the \code{\link[base]{scale}} function in that it centers 
and scales data, but scaling is accomplished by dividing the centered data by two standard 
deviations rather than one as detailed in Gelman (2008).
}
\section{References}{

Gelman, A. (2008). Scaling regression inputs by dividing by two standard deviations. 
Statistics in Medicine, 27(15), 2865–2873. \url{https://doi.org/10.1002/sim.3107}
}

