% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nocturnal.R
\name{theme_nocturnal}
\alias{theme_nocturnal}
\title{A ggplot2 theme that's dark as the night, and easy on the eyes}
\usage{
theme_nocturnal()
}
\description{
This plot theme was inspired by a need to simultaneously plot data and maintain a reasonable level of dark-adapted (scotopic) visual ability while doing fieldwork at night.
\code{theme_nocturnal()} can be used in the same way as any of the native ggplot2 themes.
}
\section{Details}{

To make adjustments to your plot's theme while using this function, using ggplot2's
\code{\link[ggplot2]{theme()}} in a line proceeding \code{theme_nocturnal()}.
}

\examples{
#Creating a plot with theme_nocturnal

DarkPlot <- ggplot(cars, aes(speed, dist)) +
  geom_point(color = "grey80") + #Note the specification of a light color for the points
  theme_nocturnal()

#Changing the axis text size of a theme_nocturnal plot

DarkPlot2 <- DarkPlot +
  theme(axis.text = element_text(size = 4))

}
