% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectsPlotter.R
\name{EffectsPlotter}
\alias{EffectsPlotter}
\title{Visualize model effects}
\usage{
EffectsPlotter(
  TopMods,
  ParamLabs = NULL,
  ConfInts = c(90, 80),
  Scaled = T,
  ThemeBlack = T
)
}
\arguments{
\item{TopMods}{This argument can accept multiple input types, and is used to specify which models effects plots
should be generated for. If an effects plot is desired for a single model, users can supply the unquoted model
object name. If effects plots for multiple models are desired, users can specify a chahracter vector of model names.}

\item{ParamLabs}{A character vector of labels for each model estimate. Labels must be supplied for estimates in the
same order as the vector returned by \code{row.names(summary(x)$coeff$cond)}, where \code{x} is a model
object name.}

\item{ConfInts}{A numeric vector specifying up to three condifence intervals as percent values from largest
to smallest (e.g., for 80\%, 90\%, and 95\% confidence intervals, supply \code{ConfInts = c(95, 90, 80)}).}

\item{Scaled}{A logical value indicating whether model estimates are scaled (T) or not (F). When \code{Scaled = T},
there will be a dotted \code{\link[ggplot2]{geom_abline}} where estimates = 0.}

\item{ThemeBlack}{A logical value indicating whether \code{\link{theme_nocturnal}} should be applied to plots
(T) or not (F).}
}
\value{
This function generates effects plots using \code{\link[ggplot2]{geom_point}} for point estimates and
\code{\link[ggplot2]{geom_crossbar}} for confidence intervals. Resultant plots are named with the model name
followed by "EffectsPlot" (e.g., for a model named "EpfuNb2", the effects plot would be named "EpfuNb2EffectsPlot").
}
\description{
This function produces effects plots for the conditional model of model objects produced by
\code{\link[glmmTMB]{glmmTMB}}. Effects can be plotted for multiple models in a single function call.
These effects plots are based on those found in 
\href{https://groups.nceas.ucsb.edu/non-linear-modeling/projects/owls/WRITEUP/owls.pdf/@@download}{this document.}
}
\section{Details}{

Because all models specified in a call to this function share estimate labels, all models in a a given call
to this function must have the same model formula.
}

\examples{
data("Epfu_Nb2_Long", "Myev_Nb2_Long", package = "EcoCountHelper")
Labels <- letters[1:12]

#Effects plot for a single model

EffectsPlotter(Epfu_Nb2_Long, Labels)
Epfu_Nb2_LongEffectsPlot

#Effects plot for multiple models specified
#in a character vector.

Mods <- c("Epfu_Nb2_Long", "Myev_Nb2_Long")
EffectsPlotter(Mods, Labels)
Epfu_Nb2_LongEffectsPlot
Myev_Nb2_LongEffectsPlot

#Effects plot for multiple models specified
#in a data frame

ModTable <- data.frame(Species = c("Epfu", "Myev"),
                        Mods = c("Epfu_Nb2_Long", "Myev_Nb2_Long"))
                        
EffectsPlotter(ModTable$Mods, Labels)

Epfu_Nb2_LongEffectsPlot
Myev_Nb2_LongEffectsPlot

}
