% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResidPlotDocs.R, R/ResidPlotLong.R,
%   R/ResidPlotWide.R
\name{ResidPlotLong}
\alias{ResidPlotLong}
\alias{ResidPlotWide}
\title{Efficiently examine multiple model fits}
\usage{
ResidPlotLong(
  Data,
  CountCol,
  GroupCol,
  ModNames,
  GroupPat = "^[[:alnum:]]+",
  Looped = T,
  Nsims = 1000,
  TestVals = T
)

ResidPlotWide(
  Data,
  ModNames,
  GroupPat = "^[[:alnum:]]+",
  Nsims = 1000,
  TestVals = T
)
}
\arguments{
\item{Data}{The unquoted name of the data frame or data table containing the raw data models are based on.}

\item{CountCol}{A character string specifying the name of the vector containing count data.}

\item{GroupCol}{A character string specifying the name of the vector containing group information (e.g., order, species) for each observation.}

\item{ModNames}{A character vector of model object names to assess model fit for.}

\item{GroupPat}{A quoted regular expression that can be used to extract the character string representing
the column name containing the response variable of a model for (e.g., if the models "EpfuNb2" and "MyluNb2" are supplied for
the \code{ModNames} argument, an appropriate regular expression would be "^[[:alpha:]]\{4\}" because the response
variable columns are "Epfu" and "Mylu", respectively).}

\item{Looped}{A logical value indicating whether models were generated using a for loop or
apply function that results in the "Data" value in the model output including an iterative subset. 
Defaults to "T".}

\item{Nsims}{A number indicating the number of data simulations to run with \code{\link[stats]{simulate}}.}

\item{TestVals}{A logical value indicating whether model fit test values should be generated alongside
the plots this function creates.}
}
\value{
These functions output the plots generated by \code{\link[DHARMa]{testResiduals}}. Each resultant plot is named with
the model name followed by "SimResidPlot" (e.g., the plot associated with a model named "EpfuNb2" would be named
"EpfuNb2SimResidPlot"). If \code{TestVals = T}, test values will also be generated. The resultant test values
are named with the model name follow by "ResidTests" (e.g., the test values associated with a model named
"EpfuNb2" would be named "EpfuNb2SimResidPlot").
}
\description{
These functions are essentially wrappers for a few \code{\link[DHARMa]{DHARMa}} functions that allow users to assess model
fit for models produced using \code{\link[glmmTMB]{glmmTMB}} with or without random effects.
}
\examples{
#ResidPlotWide Example
data("BatDataWide", "EpfuNb2Wide", "MyevNb2Wide", package = "EcoCountHelper")

ResidPlotWide(BatDataWide, c("EpfuNb2Wide", "MyevNb2Wide"), "^[[:alpha:]]{4}")

EpfuNb2WideResidTests
EpfuNb2WideSimResidPlot
MyevNb2WideResidTests
MyevNb2WideSimResidPlot

#ResidPlotLong Example
data("BatDataLong", "EpfuNb2Long", "MyevNb2Long", package = "EcoCountHelper")

ResidPlotLong(BatDataLong, "Count", "Species",
 c("EpfuNb2Long", "MyevNb2Long"), "^[[:alpha:]]{4}")

EpfuNb2LongResidTests
EpfuNb2LongSimResidPlot
MyevNb2LongResidTests
MyevNb2LongSimResidPlot

}
